buggy

+++ nanos-lite/src/ramdisk.c
@@ -23,7 +23,7 @@ size_t ramdisk_write(const void *buf, size_t offset, size_t len) {
 
 void init_ramdisk() {
   Log("ramdisk info: start = %p, end = %p, size = %d bytes",
-      &ramdisk_start, &ramdisk_end, RAMDISK_SIZE);
+      &ramdisk_start, &ramdisk_end, (int)RAMDISK_SIZE);
 }
 
 size_t get_ramdisk_size() {
